Refactored metadata reader code
